[UI] Pixel tweak
We have some rounding error problems. When we scale the width, we get
floating numbers which are not very good for our own calculations.

By reducing the margin allowance we seem to adjust a bit better the
width but it might be that two perfect sized images will overflow due
rounding errors.

Not sure how to fix this.

Signed-off-by: Sergio A. Martinez <sergio@engineer.blue>