[CI] Reimport current_user
When the client is created, it seems that it doesn't create the
current_user, or it is being modified on my manual login.

This patch fixes the missing current_user class during testing.
Some investigation has to happen to figure out why it is not importing
the right current_user

Signed-off-by: Sergio A. Martinez <sergio@engineer.blue>