[Subscriptions] Added incremental messages to subs
We send a summary of the news and recommendations.
Currently the IA hallucinates quite a bit about the state of the portfolio, due not having information of the amount of shares, when did you joing a company etc.

It is quite interesting but not sure how a paying user would take the halluciations

As a piece of entertaining is quite good.

Signed-off-by: Sergio A. Martinez <sergio@engineer.blue>